# Cache the index of the last accessed polynomial